#11 fix : package 구조 변경
