--- src/main/java/domain/enums/Gender.java
@@ -1,5 +1,0 @@
-package domain.enums;
-
-public enum Gender {
-    MAN, WOMAN
-}
