+++ src/main/java/ticketing/ticketing/domain/entity/Cast.java
@@ -1,0 +1,32 @@
+package ticketing.ticketing.domain.entity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.springframework.boot.autoconfigure.kafka.KafkaProperties.Admin;
+
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)  // JPA 기본 생성자 (리플렉션을 위해 필수)
+@AllArgsConstructor(access = AccessLevel.PRIVATE)   // Builder 내부에서만 사용
+@Builder(access = AccessLevel.PROTECTED)            // 외부에서 builder 직접 사용 불가
+@Getter
+public class Cast {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String name;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "admin_id")
+    private Admin admin;
+}
