+++ src/main/java/ticketing/ticketing/domain/entity/ConcertSeatMetadata.java
@@ -1,9 +1,6 @@
-package domain.entity;
+package ticketing.ticketing.domain.entity;
 
-import domain.enums.ReportReason;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -11,35 +8,31 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import java.time.LocalDateTime;
+import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)  // JPA 기본 생성자 (리플렉션을 위해 필수)
+@AllArgsConstructor(access = AccessLevel.PRIVATE)   // Builder 내부에서만 사용
+@Builder(access = AccessLevel.PROTECTED)            // 외부에서 builder 직접 사용 불가
 @Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class Report {
+public class ConcertSeatMetadata {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "admin_id")
-    private Admin admin;
-
-    @Enumerated(EnumType.STRING)
-    private ReportReason reason;
+    @JoinColumn(name = "seats_id")
+    private Seats seats;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "review_id")
-    private Review review;
+    @JoinColumn(name = "concert_id")
+    private Concert concert;
 
+    private Integer price;
+    private Long seatPerAblePerson;
     private LocalDateTime createdAt;
-    private LocalDateTime updatedAt;
-    private LocalDateTime deletedAt;
 }
