+++ src/main/java/ticketing/ticketing/domain/entity/Reservation.java
@@ -1,7 +1,5 @@
-package domain.entity;
+package ticketing.ticketing.domain.entity;
 
-import domain.enums.InquiryStatus;
-import domain.enums.InquiryType;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
@@ -10,22 +8,20 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
-import jakarta.persistence.Lob;
 import jakarta.persistence.ManyToOne;
 import java.time.LocalDateTime;
+import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)  // JPA 기본 생성자 (리플렉션을 위해 필수)
+@AllArgsConstructor(access = AccessLevel.PRIVATE)   // Builder 내부에서만 사용
+@Builder(access = AccessLevel.PROTECTED)            // 외부에서 builder 직접 사용 불가
 @Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class Inquiry {
+public class Reservation {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
@@ -34,19 +30,18 @@ public class Inquiry {
     @JoinColumn(name = "user_id")
     private User user;
 
-    private String title;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "concert_id")
+    private Concert concert;
 
-    @Lob
-    private String content;
-
-    @Enumerated(EnumType.STRING)
-    private InquiryType type;
+    private Long ticketCount;
 
     @Enumerated(EnumType.STRING)
-    private InquiryStatus status;
+    private ReservationState state;
 
-    private LocalDateTime repliedAt;
     private LocalDateTime createdAt;
-    private LocalDateTime updatedAt;
-    private LocalDateTime deletedAt;
+
+    public enum ReservationState {
+        RESERVED, CANCELLED
+    }
 }
