+++ src/main/java/ticketing/ticketing/domain/entity/User.java
@@ -1,28 +1,25 @@
-package domain.entity;
+package ticketing.ticketing.domain.entity;
 
-import domain.enums.Gender;
-import domain.enums.UserState;
+import ticketing.ticketing.domain.enums.Gender;
+import ticketing.ticketing.domain.enums.UserState;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.Table;
 import java.time.LocalDateTime;
+import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @Entity
-@Table(name = "users")
+@NoArgsConstructor(access = AccessLevel.PROTECTED)  // JPA 기본 생성자 (리플렉션을 위해 필수)
+@AllArgsConstructor(access = AccessLevel.PRIVATE)   // Builder 내부에서만 사용
+@Builder(access = AccessLevel.PROTECTED)            // 외부에서 builder 직접 사용 불가
 @Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
