+++ src/main/java/ticketing/ticketing/domain/enums/AdminRole.java
@@ -1,4 +1,4 @@
-package domain.enums;
+package ticketing.ticketing.domain.enums;
 
 public enum AdminRole {
     CONCERT_ADMIN, SITE_ADMIN
