+++ src/main/java/ticketing/ticketing/domain/enums/CastRole.java
@@ -1,4 +1,4 @@
-package domain.enums;
+package ticketing.ticketing.domain.enums;
 
 public enum CastRole {
     MAIN, SUPPORT, GUEST
