+++ src/main/java/ticketing/ticketing/domain/enums/Gender.java
@@ -1,0 +1,5 @@
+package ticketing.ticketing.domain.enums;
+
+public enum Gender {
+    MAN, WOMAN
+}
