+++ src/main/java/ticketing/ticketing/domain/enums/InquiryStatus.java
@@ -1,4 +1,4 @@
-package domain.enums;
+package ticketing.ticketing.domain.enums;
 
 public enum InquiryStatus {
     PENDING, COMPLETED
