+++ src/main/java/ticketing/ticketing/domain/enums/InquiryType.java
@@ -1,4 +1,4 @@
-package domain.enums;
+package ticketing.ticketing.domain.enums;
 
 public enum InquiryType {
   RESERVATION, PRODUCT, DELIVERY, CANCELLATION, USER_MANAGEMENT, ETC
