+++ src/main/java/ticketing/ticketing/domain/enums/ReportReason.java
@@ -1,4 +1,4 @@
-package domain.enums;
+package ticketing.ticketing.domain.enums;
 
 public enum ReportReason {
     ABUSE, SPAM, INAPPROPRIATE, ETC
