+++ src/main/java/ticketing/ticketing/domain/enums/ReservationState.java
@@ -1,4 +1,4 @@
-package domain.enums;
+package ticketing.ticketing.domain.enums;
 
 public enum ReservationState {
     RESERVED, CANCELLED
