+++ src/main/java/ticketing/ticketing/domain/enums/UserState.java
@@ -1,4 +1,4 @@
-package domain.enums;
+package ticketing.ticketing.domain.enums;
 
 public enum UserState {
     ACTIVE, INACTIVE    // 계정 정지에 관한 Enum
